options/internal: Switch to frg::sharded_slab
